# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AWSChatBot",
    "version": "1.0.0"
  },
  "paths": {
    "/users": {
      "get": {
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "List of users"
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created"
          }
        }
      }
    },
    "/users/{emailID}": {
      "put": {
        "summary": "Update a user by emailID",
        "parameters": [
          {
            "name": "emailID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "emailID": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "conversationHistory": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "fullName",
          "emailID",
          "phoneNumber"
        ]
      }
    }
  }
}